settings to its own component (#9)
* settings to its own component

* getsize to switch of maps

* Change "Save Settings" to "Apply"

"Apply" is more succinct, and makes more sense in a process where the changes are only made once you press the button.  If the page size was changing at the same time as each field was changed, then "Save Settings" would make more sense.

* Add HR element to separate Print to PDF

The Export feature isn't really a part of "sheet settings", and might be suited better to another area of the screen entirely.  But for now, it can live here in the settings, but it should be separated from the other Settings buttons.

* disable Apply/Save button until changes made.

Two things in this:  first, the Apply button should be enabled only when changes have been made to any settings.  second, the settings fields weren't accurately being loaded with the page settings (when loading from localStorage).  I believe this is because the inputs need to be 'controlled' inputs.

* Change "Save Settings" to "Apply"

"Apply" is more succinct, and makes more sense in a process where the changes are only made once you press the button.  If the page size was changing at the same time as each field was changed, then "Save Settings" would make more sense.

* add onChange to rowHeight input

---------
